move audio observation logic into VideoFeedAudioController
